drf-api/pyproject.toml
[tool.poetry]
name = "Api"
version = "0.1.0"
description = ""
authors = ["ahmadjaara <ahmadjaara98@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
